sm6150-common: add wifi mac generator
The default qcacld-3.0 implementation looks for the MAC
in a firmware with the name `wlan/qca_cld/wlan_mac.bin`.
That firmware is, by default, symlinked to
`/mnt/vendor/persist/wlan_mac.bin`.
Xiaomi has somewhat changed the qcacld-3.0 logic to load
the MAC from another place. Xiaomi's nvmac puts a plain-text
MAC in `/data/vendor/mac_addr/wlan_mac.bin`.
Add a script to write the MAC address found in the file
created by nvmac to the file in persist that the firmware
file is symlinked to.

This implementation is (loosely) based on the one
written for OnePlus 7.

There were a few problems with the old code.
 * If the last character of the MAC was a letter, it wouldn't
   be incremented because we were replacing lowercase letters
   and not uppercase letters using tr, and the MAC only had
   uppercase letters.
 * If the last character of the MAC was an F, it would be replaced
   with an A rather than a 0.
 * If the last character of the MAC was an F, it will get replaced,
   but the character before the last wouldn't change (and so on).

Fix all these problems by using decimal incrementation.

Change-Id: I8e22ddbc8a8b6b5f493a32ab179e48c1815e19bc